refactor: . pull addTask(), addEditButton(), addUrgency() up
